Cleaned up SerialDevice, Device, DeviceManager, implemented waiting for core devices
