--- FusorControlServer/src/main/java/com/eastsideprep/fusorcontrolserver/Device.java
@@ -1,11 +1,0 @@
-package com.eastsideprep.fusorcontrolserver;
-
-public abstract class Device {
-
-    public abstract void write(String s);
-
-    public boolean isControl() {
-        // by default no. need to override in subclass
-        return false;
-    }
-}
