+++ FusorControlServer/src/main/java/com/eastsideprep/fusorcontrolserver/WebServer.java
@@ -99,7 +99,7 @@ public class WebServer {
         long start = System.currentTimeMillis();
         String tmpv = "";
         String tmps = "";
-        dm.updateAllStatus();
+        //dm.updateAllStatus(); // todo redesign and reimplement
         String message = "hold";
         List<Status> status = new ArrayList<>();
